+++ .idea/libraries/Gson.xml
@@ -3,6 +3,8 @@
     <properties maven-id="com.google.code.gson:gson:LATEST" />
     <CLASSES>
       <root url="jar://$MAVEN_REPOSITORY$/com/google/code/gson/gson/2.5/gson-2.5.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/com/google/code/gson/gson/2.5/gson-2.5.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/com/google/code/gson/gson/2.5/gson-2.5.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
