+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'TheBlueAlliance'
+rootProject.name = 'thebluealliance'
